Fixed bugs in GetUserTransaction
